Move `MagicCommentParser` under `FileParser` namespace
