+++ lib/annotate_rb/model_annotator.rb
@@ -22,10 +22,10 @@ module AnnotateRb
     autoload :SingleFileRemoveAnnotationInstruction, "annotate_rb/model_annotator/single_file_remove_annotation_instruction"
     autoload :AnnotationDiffGenerator, "annotate_rb/model_annotator/annotation_diff_generator"
     autoload :AnnotationDiff, "annotate_rb/model_annotator/annotation_diff"
-    autoload :MagicCommentParser, "annotate_rb/model_annotator/magic_comment_parser"
     autoload :FileComponents, "annotate_rb/model_annotator/file_components"
     autoload :ProjectAnnotator, "annotate_rb/model_annotator/project_annotator"
     autoload :ProjectAnnotationRemover, "annotate_rb/model_annotator/project_annotation_remover"
     autoload :AnnotatedFile, "annotate_rb/model_annotator/annotated_file"
+    autoload :FileParser, "annotate_rb/model_annotator/file_parser"
   end
 end
