+++ lib/annotate_rb/model_annotator/file_parser.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module AnnotateRb
+  module ModelAnnotator
+    module FileParser
+      autoload :MagicCommentParser, "annotate_rb/model_annotator/file_parser/magic_comment_parser"
+    end
+  end
+end
