+++ spec/lib/annotate_rb/model_annotator/file_parser/magic_comment_parser_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe AnnotateRb::ModelAnnotator::MagicCommentParser do
+RSpec.describe AnnotateRb::ModelAnnotator::FileParser::MagicCommentParser do
   describe ".call" do
     subject { described_class.call(content) }
